requirements.txt
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.2.0
DateTime==5.2
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
python-dotenv==1.0.0
python-telegram-bot==20.4
pytz==2023.3
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
urllib3==2.0.4
zope.interface==6.0
